refactored to move code to seperate files and packages
