+++ Reply/app/src/main/res/values/strings.xml
@@ -16,5 +16,5 @@
     <string name="compose">Compose</string>
 
     <string name="empty_screen_title">Screen under construction</string>
-    <string name="empty_screen_subtitle">This screen is still under construction. This sample will help you learn about adaptive layouts in Jetpack compose</string>
+    <string name="empty_screen_subtitle">This screen is still under construction. This sample will help you learn about adaptive layouts in Jetpack Compose</string>
 </resources>
